s31890/08/requirements.txt
scikit-learn
matplotlib
numpy
pandas
imbalanced-learn
tensorflow
keras-tuner
tensorflow-datasets
pillow
seaborn
